actualizar searchbar
