+++ src/App.jsx
@@ -1,15 +1,13 @@
-import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
 import './App.css'
-import Toys from './Products'
+import SearchBar from './components/SearchBar'
+import Toys from './components/Toys'
 
 function App() {
-  const [count, setCount] = useState(0)
 
   return (
     <>
     <div>
+      <SearchBar />
      <Toys />
      </div>
     </>
