+++ src/components/SearchBar.jsx
@@ -1,0 +1,20 @@
+import React from "react"
+
+export default function SearchBar() {
+
+    return (
+
+        <div>
+            <div class="relative bg-gray-900 p-2 rounded border-2 border-blue-300">
+                <svg class="h-5 w-5 absolute left-0 ml-2" fill="currentColor" viewBox="0 0 20 20">
+                    <path fill-rule="evenodd"
+                        d="M8 4a4 4 0 100 8 4 4 0 000-8zM2 8a6 6 0 1110.89 3.476l4.817 4.817a1 1 0 01-1.414 1.414l-4.816-4.816A6 6 0 012 8z"
+                        clip-rule="evenodd">
+                    </path>
+                </svg>
+                <input type="text" placeholder="Buscar..." class="ml-6 bg-transparent border-opacity-0" />
+            </div>
+        </div>
+
+    )
+}
